Create o modulo de um número negativo
Write a small program which prints the absolute value of a given number if the number is less than zero. Next it should print the value of the given number on a separate line.

Your program must print the same result as the expected output. Test it for several other cases (positive numbers, other negative numbers, etc.)