Create pi and day of the week ( rounding floats )
Write a program which takes two values: a count of the days in one week and the value of Pi. Next, print these two values. Don't forget about data types.

Your version of the program must print the same result as the expected output.

Sample Input

7
3.14
Output

Sample Output

How many days in the week: 7
The value of Pi to two points: 3.14