Create day of the week
Write a program that prints the name of a given day of the week. Your program must print the same result as the expected output. This task is similar to the previous lab, but this time you have to get the day of the week from the user and validate the input.

Test the program for all the days of the week (add code to print a message to the user when he/she enters an invalid day of the week).